Refactoring custom hook to make it more general
