+++ src/hooks/useRequestDelay.js
@@ -1,0 +1,56 @@
+import { useState, useEffect } from "react";
+
+export const REQUEST_STATUS = {
+    LOADING: 1,
+    SUCCESS: 2,
+    FAILURE: 3
+};
+
+function useRequestDelay(delayTime = 1000, initialData = []) {
+    const [data, setData] = useState(initialData);
+    const [requestStatus, setRequestStatus] = useState(REQUEST_STATUS.LOADING);
+    const [error, setError] = useState("");
+
+
+    const delay = (ms) => new Promise((resolve) => setTimeout(resolve, ms));
+
+    useEffect(() => {
+        async function delayFunc() {
+            try {
+                await delay(delayTime);
+                // throw "Had error";
+                setRequestStatus(REQUEST_STATUS.SUCCESS);
+                setData(data);
+            }
+            catch (e) {
+                setRequestStatus(REQUEST_STATUS.FAILURE)
+                setError(e);
+            }
+        }
+
+        delayFunc();
+    }, []);
+
+    function updateRecord(recordUpdated) {
+        
+            const newRecords = data.map(function (rec) {
+                return rec.id === recordUpdated.id ? recordUpdated : rec;
+            });
+            
+        async function delayFunc() {
+            try {
+                await delay(delayTime);
+                setData(newRecords);
+            }
+            catch (e) {
+               console.log("Error loading", e);
+            }
+        }
+
+        delayFunc();
+    }
+
+    return { data, requestStatus, error, updateRecord };
+}
+
+export default useRequestDelay;
